Pull stuff out of activator
